==> 02_dlbcl/04_computeScores.Rmd <==
---
title: "Pathway signals"
author: "Sergio Picart-Armada"
date: "September 12, 2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Introduction

## Load libraries

```{r}
library(igraph)
library(diffuStats)

library(digest)
library(tictoc)

library(plyr)
library(dplyr)
library(magrittr)

library(emmeans)

library(ggplot2)
library(ggsci)

source("params.R")
files_signal <- list.files(dir_simul)
stopifnot(length(files_signal) > 0)

dir.create(dir_scores)
dir.create(dir_models)
dir.create(dir_stats)

```

## Network and kernel

```{r}
# basic files
load(file_kernel)
load(file_graph)
bkgd <- V(g_cc)$geneID
stopifnot(length(bkgd) == length(unique(bkgd)))

# pagerank
pr <- setNames(
  page.rank(graph = g_cc, directed = FALSE)$vector, 
  V(g_cc)$geneID
) 

# gene_array <- list_arrays[1]
# file_signal <- files_signal[1]
# strat <- list_strat[1]
```

# Diffusion scores and metrics

```{r, message=FALSE, warning=FALSE}
tic("Heavy computations")
plyr::l_ply(
  list_arrays, 
  function(gene_array) {
    # get the array genes
    attrib <- get.vertex.attribute(g_cc, gene_array)
    
    # ids, must have names in list_strat
    id_obs <- V(g_cc)[attrib]$geneID
    id_hid <- V(g_cc)[!attrib]$geneID
    id_all <- V(g_cc)$geneID
    
    # create folder
    path_scores_array <- paste0(dir_scores, "/", gene_array)
    dir.create(path_scores_array)
    path_stats_array <- paste0(dir_stats, "/", gene_array)
    dir.create(path_stats_array)
    # path_plots_array <- paste0(dir_plots, "/", gene_array)
    # dir.create(path_plots_array)
    
    # iterate over the artificial inputs
    plyr::l_ply(
      files_signal, 
      function(file_signal) {
        # load generated input (varname = signal)
        file.load <- paste0(dir_simul, "/", file_signal)
        load(file = file.load)
        
        # iterate over FDRs
        plyr::l_ply(
          list_fdr, 
          function(threshold_fdr) {
            
            # Binarise input 
            input <- apply(
              signal$mat_input[id_obs, ], 
              2, 
              function(x) {
                1*(p.adjust(x, method = "fdr") < threshold_fdr)
              })
            
            # make sure there is at least one column 
            # with some non-zero data
            non_empty <- colSums(input) > 0
            if (sum(non_empty) < 2) return(NULL)
    
            # message(non_empty, " nonempty cols")
            input <- input[, non_empty, drop = FALSE]
            
            # diffusion scores
            list_diff <- lapply(
              setNames(list_methods, list_methods), 
              function(m) {
                diffuStats::diffuse(
                  K = K, scores = input, method = m, n.perm = n.perm
                )
              }
            )
            
            # baseline: original p-values as a ranker
            list_diff$original <- -log(signal$mat_input[, non_empty])
            # baseline: centrality values (pagerank)
            list_diff$pagerank <- replicate(n = ncol(input), expr = pr) %>% 
                set_colnames(colnames(input))
            
            # dummy random scores
            # See note on the same procedure in the cross validation
            # place random seed from hasing
            seed.index <- digest::digest(file.load) %>% substr(1, 5) %>% strtoi(base = 16)
            set.seed(seed.index)
            list_diff$random <- replicate(n = ncol(input), expr = sample(length(pr))) 
            dimnames(list_diff$random) <- dimnames(list_diff$pagerank) 
    
            # these files are heavy and is saved in their own directory (git-ignored)
            save(list_diff, 
                 file = paste0(path_scores_array, "/fdr_", threshold_fdr, ",", file_signal), 
                 compress = "xz")
            
            # validate in each background: obs, hid, all
            plyr::l_ply(
              list_strat, 
              function(strat) {
                # browser()
                # create folderpath_scores_array
                path_stats_array_strat <- paste0(path_stats_array, "/", strat)
                dir.create(path_stats_array_strat)
                
                # vertices in that stratification
                id_v <- get(strat)
                
                # compute metrics
                df_auc <- plyr::ldply(
                  list_diff, 
                  function(elem) {
                    # this evaluates by using the auc (default)
                    # the auc is not automatically reversed (1-auc)
                    # i.e. 1s are expected to have higher 
                    # predicted values
                    diffuStats::perf_eval(
                      prediction = elem[id_v, ], 
                      validation = signal$mat_validation[id_v, non_empty], 
                      metric = list(auroc = metric_fun(curve = "ROC"), 
                                    auprc = metric_fun(curve = "PRC"))
                    )
                  }, 
                  .id = "method"
                )
                mat_auc <- reshape2::acast(
                  df_auc, Column ~ method, value.var = "auroc")
                
                # wilcox test
                method_cols <- colnames(mat_auc)
                n_cols <- length(method_cols)
                
                mat_wilcox <- matrix(
                  data = NA_character_, nrow = n_cols, ncol = n_cols, 
                  dimnames = list(method_cols, method_cols))
                
                # run the tests and fill the matrix
                # for's are fast as there are only few iterations
                grid_method <- expand.grid(
                  seq_along(method_cols), 
                  seq_along(method_cols))
                
                for (r in 1:nrow(grid_method)) {
                  met1 <- grid_method[r, 1]
                  met2 <- grid_method[r, 2]
                  
                  # test method1-method2 but not the reciprocal
                  # method2-method1
                  if (met1 < met2) {
                    wil_test <- wilcox.test(
                      x = mat_auc[, met1], 
                      y = mat_auc[, met2], 
                      alternative = "two.sided", 
                      paired = TRUE, 
                      conf.int = TRUE, 
                      conf.level = 0.95
                    )
                    # conf_int
                    est <- wil_test$estimate
                    if (is.na(est)) {
                      cell <- NA_character_
                    } else {
                      cell <- paste0(
                        format(est, digits = 2), 
                        "[", 
                        format(wil_test$conf.int[1], digits = 2), 
                        ",",
                        format(wil_test$conf.int[2], digits = 2), 
                        "]"
                      )
                    }
                    mat_wilcox[met1, met2] <- cell
                    
                    # pvalue (raw)
                    p <- wil_test$p.value
                    if (is.na(p)) p <- NA_character_ # this switches NaN to NA
                    mat_wilcox[met2, met1] <- p
                  }
                }
                
                # FDR correction for p-values
                pvals <- p.adjust(
                  as.numeric(mat_wilcox[lower.tri(mat_wilcox)]), 
                  method = "fdr") 
                mat_wilcox[lower.tri(mat_wilcox)] <- format(pvals, digits = 3)
                
                # correlations
                mat_cor_auc <- cor(x = mat_auc, method = "spearman")
                
                # save metrics
                save(df_auc, mat_wilcox, mat_cor_auc, 
                     file = paste0(path_stats_array_strat, 
                                   "/fdr_", threshold_fdr, ",", file_signal))
              }
            )
          }
        )
      }
    )
  }, 
  .progress = "text"
)
toc()
```

# Added a posteriori: add AUPRC and AUROC models

```{r}
df_stats <- plyr::ldply(
  list_arrays, 
  function(gene_array) {
    # browser()
    
    # get the array genes
    attrib <- get.vertex.attribute(g_cc, gene_array)
    
    # ids, must have names in list_strat
    id_obs <- V(g_cc)[attrib]$geneID
    id_hid <- V(g_cc)[!attrib]$geneID
    id_all <- V(g_cc)$geneID
    
    # create folder
    path_scores_array <- paste0(dir_scores, "/", gene_array)
    dir.create(path_scores_array)
    path_stats_array <- paste0(dir_stats, "/", gene_array)
    dir.create(path_stats_array)
    # path_plots_array <- paste0(dir_plots, "/", gene_array)
    # dir.create(path_plots_array)
    
    # iterate over the artificial inputs
    plyr::ldply(
      setNames(files_signal, files_signal), 
      function(file_signal) {
        # browser()
        
        # load generated input (varname = signal)
        load(file = paste0(dir_simul, "/", file_signal))
        
        # iterate over FDRs
        plyr::ldply(
          setNames(list_fdr, list_fdr), 
          function(threshold_fdr) {
        
            # Binarise input 
            input <- apply(
              signal$mat_input[id_obs, ], 
              2, 
              function(x) {
                1*(p.adjust(x, method = "fdr") < threshold_fdr)
              })
            
            # make sure there is at least one column 
            # with some non-zero data
            non_empty <- colSums(input) > 0
            if (sum(non_empty) < 2) return(NULL)
    
            input <- input[, non_empty, drop = FALSE]
            
            # Load the scores
            load(paste0(path_scores_array, "/fdr_", threshold_fdr, ",", file_signal))
            
            # validate in each background: obs, hid, all
            plyr::ldply(
              setNames(list_strat, list_strat), 
              function(strat) {
                # browser()
                # create folderpath_scores_array
                # vertices in that stratification
                id_v <- get(strat)
                
                # compute metrics
                plyr::ldply(
                  list_diff, 
                  function(elem) {
                    # this evaluates by using the auc (default)
                    # the auc is not automatically reversed (1-auc)
                    # i.e. 1s are expected to have higher 
                    # predicted values
                    diffuStats::perf_eval(
                      prediction = elem[id_v, ], 
                      validation = signal$mat_validation[id_v, non_empty], 
                      metric = list(auroc = metric_fun(curve = "ROC"), 
                                    auprc = metric_fun(curve = "PRC"))
                    )
                  }, 
                  .id = "method"
                )
              }, 
              .id = "strat"
            )
          }, 
          .id = "fdr"
        )
      }, 
      .id = "file"
    )
  }, 
  .id = "array", 
  .progress = "text"
)
save(df_stats, file = paste0(dir_models, "/df_all.RData"), compress = "xz")
```

## Data frame to fit models

```{r}
# Un-paste parameter columns
cols_param <- gsub(pattern = "\\.RData$", replacement = "", x = df_stats$file) %>% 
  strsplit(split = ",") %>% 
  plyr::ldply(function(df) {
    tmp <- strsplit(df, split = "_")
    colname <- sapply(tmp, function(x) x[[1]])
    colvalues <- sapply(tmp, function(x) x[[2]])
    
    setNames(colvalues, colname)
  }) %>% 
  lapply(function(col) {
    ans <- as.factor(col)
    lev <- levels(ans)
    # sort levels increasingly
    factor(ans, levels = lev[order(as.numeric(as.character(lev)))])
  })

# new data frame for models
df_mod <- cbind(df_stats, cols_param) %>% select(-file) 
df_mod$pmax <- as.factor(format(as.numeric(as.character(df_mod$pmax)), scientific = TRUE))
save(df_mod, file = paste0(dir_models, "/df_mod.RData"), compress = "xz")
```

## Logistic models

```{r}
# Build logistic models
plyr::d_ply(
  df_mod, 
  c("array", "strat"), 
  function(df) {
    array <- df$array[1]
    strat <- df$strat[1]
    
    df <- filter(df, pmax != "1e-05" & k != "10")
    out_suffix <- paste0(",array_", array, 
                         ",", strat, 
                         ".txt")
    
    mod_logistic <- glm(
      auroc ~ method + k + r + pmax + fdr, 
      data = df, 
      family = quasibinomial(link = "logit"))
    # summary(mod_logistic)
    # anova(mod_logistic) 
    writeLines(
      capture.output(summary(mod_logistic)), 
      con = paste0(dir_models, "/auroc", out_suffix)
    )
    
    mod_logistic <- glm(
      auprc ~ method + k + r + pmax + fdr, 
      data = df, 
      family = quasibinomial(link = "logit"))
    # summary(mod_logistic)
    # anova(mod_logistic) 
    writeLines(
      capture.output(summary(mod_logistic)), 
      con = paste0(dir_models, "/auprc", out_suffix)
    )
  }
)
```

## One big model for AUPRC

```{r}
df_bigmod <- filter(df_mod, pmax != "1e-05" & k != "10")

mod_big_auprc <- glm(
  auprc ~ method + method:strat + k + r + pmax + array + fdr, 
  data = filter(df_mod, pmax != "1e-05" & k != "10"), 
  family = quasibinomial(link = "logit"))

summary_big_auprc <- summary(mod_big_auprc)

writeLines(
  capture.output(print(summary_big_auprc)), 
  con = paste0(dir_models, "/big_auprc_summary.txt")
)
write.csv(
  summary_big_auprc$coefficients, 
  file = paste0(dir_models, "/big_auprc_coef.csv"), 
  row.names = FALSE)

df_emmeans <- emmeans(
  mod_big_auprc, specs = c("method", "strat", "array"), 
  type = "response", nesting = NULL) %>% summary
write.csv(
  df_emmeans, 
  file = paste0(dir_models, "/big_auprc_emmeans.csv"), 
  row.names = FALSE)
```

## Plot model predictions

```{r}
g <- ggplot(df_emmeans, aes(x = method, colour = method)) + 
  # geom_hline(yintercept = 0, lty = 3, color = "gray45") + 
  geom_errorbar(aes(ymin = asymp.LCL, ymax = asymp.UCL), width = .3, position = "dodge") + 
  facet_grid(array~strat, scales = "free", space = "free_y", drop = TRUE) + 
  scale_colour_npg(guide = FALSE) + 
  # coord_flip() +
  theme_bw() + 
  xlab("Method") + 
  ylab("Predicted mean AUPRC") + 
  theme(strip.text.y = element_text(angle = 0), 
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, colour = "black"), 
        axis.text.y = element_text(colour = "black")) 
ggsave(g, file = paste0(dir_models, "/big_auprc_predicted.png"), 
       width = 7, height = 5, units = "in")
```

# Reproducibility

```{r}
out <- capture.output(sessionInfo())
writeLines(out, con = paste0(dir_metadata, "/04_sessionInfo.txt"))
```
