==> 02_dlbcl/05_plots.Rmd <==
---
title: "Performance plots"
author: "Sergio Picart-Armada"
date: "September 12, 2017"
output: html_document
---

# Plot the statistics of the diffusion datasets

```{r}
library(plyr)
library(magrittr)

library(ggplot2)
library(ggsci)

source("params.R")
load(paste0(dir_models, "/df_mod.RData"))

dir.create(dir_plots)
```

# Plot AUROC / AUPRC

```{r}
plyr::l_ply(
  c("auprc", "auroc"), 
  function(metric) {
    # iterate over values of pmax
    plyr::d_ply(
      df_mod, 
      c("array", "strat", "pmax", "fdr"), 
      function(df) {
        # browser()
        pm <- df$pmax[1]
        arr <- df$array[1]
        strat <- df$strat[1]
        fdr <- df$fdr[1]
        
        g <- ggplot(df, aes_string(y = metric, x = "method")) + 
          # geom_hline(yintercept = .5, color = "red", alpha = .7) + 
          geom_boxplot(aes(fill = method), 
                       colour = "gray10", 
                       outlier.size = .1, 
                       lwd = .2) + 
          scale_fill_npg() + 
          facet_grid(k ~ r) + 
          xlab("Method") +
          ylab(metric)  +
          theme_bw() + 
          theme(aspect.ratio = 1, 
                axis.text.x = element_text(angle = 45, hjust = 1), 
                text = element_text(family = "arial", size = 9)) 
          # scale_y_continuous(limits = 0:1, 
          #                    expand = c(0, 0), 
          #                    breaks = 0:10/10) + 
          
        
        out_name <- paste0(
          dir_plots, "/", metric, ",array_", arr, ",", strat, ";fdr_", fdr, ",pmax_", 
          format(as.numeric(as.character(pm)), scientific = TRUE), ".png"
        )
        
        ggsave(filename = out_name, plot = g, width = 178, units = "mm")
      }
    )
  }
)
```

# Reproducibility

```{r}
out <- capture.output(sessionInfo())
writeLines(out, con = paste0(dir_metadata, "/05_sessionInfo.txt"))
```
